+++ index.ts
@@ -1,6 +1,6 @@
 import { ValidationErrorCodes } from "./src/error-codes";
 import { validateProtocolConfiguration } from "./src/validate-protocol-configuration";
-import { MetaProtocolDefinition, BuiltMetaProtocolDefinition } from "./src/meta-protocol-type";
+import { MetaProtocolDefinition, BuiltMetaProtocolDefinition } from "./src/type/meta-protocol-type";
 import { isMetaProtocol } from "./src/is-meta-protocol";
 import { ValidateProtocolClass } from "./src/validate-protocol-class";
 import { MetaProtocol } from "./src/meta-protocol";