+++ src/bin/validate-protocol.ts
@@ -6,5 +6,6 @@ export const validateProtocol = async (workingDir = "", isDbProtocol = false) : 
   console.log(processing("Starting validation of the \"meta-protocol.json\" file...\n"));
 
   const fileContent = await getDescriptorFileContent(workingDir, "meta-protocol.json");
+
   await validateProtocolConfiguration(fileContent, workingDir, isDbProtocol);
 };