+++ test/db-protocol-tooling.spec.ts
@@ -71,5 +71,102 @@ describe("DbProtocols Tooling", () => {
       expect(result["1234"].changes[0].newState).to.be.equal(null);
       expect(result["1234"].changes[0].path).to.be.equal("FULL_SCHEMA");
     });
+
+    it("Checks added Schema property Diff", () => {
+      const originalSchema : SchemaType = {
+        name: "testSchema",
+        format: {
+          property1: { "type": "string" },
+          property2: { "type": "number" },
+          property3: { "type": "boolean" },
+        },
+        dbProtocol: "",
+        identifier: "1234",
+      };
+
+      const changedSchema : SchemaType = {
+        name: "testSchema",
+        format: {
+          property1: { "type": "string" },
+          property2: { "type": "number" },
+          property3: { "type": "boolean" },
+          property4: { "type": "number" },
+        },
+        dbProtocol: "",
+        identifier: "1234",
+      };
+
+      const result = checkSchemaDiff([originalSchema], [changedSchema]);
+
+      expect(result["1234"].changes[0].action).to.be.equal("added");
+      expect(result["1234"].changes[0].newState).to.be.deep.equal({ type: "number" });
+      expect(result["1234"].changes[0].path).to.be.equal("format.property4");
+    });
+
+    it("Checks removed Schema property Diff", () => {
+      const changedSchema : SchemaType = {
+        name: "testSchema",
+        format: {
+          property1: { "type": "string" },
+          property2: { "type": "number" },
+          property3: { "type": "boolean" },
+        },
+        dbProtocol: "",
+        identifier: "1234",
+      };
+
+      const originalSchema : SchemaType = {
+        name: "testSchema",
+        format: {
+          property1: { "type": "string" },
+          property2: { "type": "number" },
+          property3: { "type": "boolean" },
+          property4: { "type": "number" },
+        },
+        dbProtocol: "",
+        identifier: "1234",
+      };
+
+      const result = checkSchemaDiff([originalSchema], [changedSchema]);
+
+      expect(result["1234"].changes[0].action).to.be.equal("removed");
+      expect(result["1234"].changes[0].newState).to.be.deep.equal(null);
+      expect(result["1234"].changes[0].path).to.be.equal("format.property4");
+    });
+
+    it("Checks changed type of Schema property Diff", () => {
+      const originalSchema : SchemaType = {
+        name: "testSchema",
+        format: {
+          property1: { "type": "string" },
+          property2: { "type": "number" },
+          property3: { "type": "boolean" },
+        },
+        dbProtocol: "",
+        identifier: "1234",
+      };
+
+      const changedSchema : SchemaType = {
+        name: "testSchema",
+        format: {
+          property1: { "type": "string" },
+          property2: { "type": "number" },
+          property3: { "type": "object", subtype: {
+            anotherProp: { "type": "string" },
+          } },
+        },
+        dbProtocol: "",
+        identifier: "1234",
+      };
+
+      const result = checkSchemaDiff([originalSchema], [changedSchema]);
+
+      expect(result["1234"].changes[0].action).to.be.equal("type_changed");
+      expect(result["1234"].changes[0].newState).to.be.deep.equal("object");
+      expect(result["1234"].changes[0].path).to.be.equal("format.property3.type");
+      expect(result["1234"].changes[1].action).to.be.equal("added");
+      expect(result["1234"].changes[1].newState).to.be.deep.equal({ anotherProp: { type: "string" } });
+      expect(result["1234"].changes[1].path).to.be.equal("format.property3.subtype");
+    });
   });
 });