+++ test/test-data/db-protocols/correct/test.js
@@ -14,6 +14,8 @@ class TestDBProtocol extends imported.DBProtocol {
 
   validateConfiguration () {}
 
+  verifySchemaSupport () {}
+
   getProtocolPublicMethods () {
     return {
       aNamedFunction: () => {},